+++ src/Auth/loginCommand.ts
@@ -4,9 +4,10 @@ import * as Debug from 'debug';
 import * as os from 'os';
 import { deserializeJSON, getApiUrl, postResource } from '../helper';
 import { saveConfig, getConfigFilePath, loadConfig } from '../RunControl/runControlHelper';
-import * as parameters from '../../config/parameters';
+import { parameters } from '../parameters';
 import { CommandLineOptions, createCommandDefinition } from '../Command/commandDefinition';
 import { ApiVersions } from '@signageos/sdk/dist/RestApi/apiVersions';
+import { log } from '@signageos/sdk/dist/Console/log';
 const debug = Debug('@signageos/cli:Auth:login');
 
 const OPTION_LIST = [
@@ -49,7 +50,7 @@ export const login = createCommandDefinition({
 			apiSecurityToken,
 		});
 
-		console.log(`User ${chalk.green(identification!)} has been logged in with token "${name}". Credentials are stored in ${chalk.blue(getConfigFilePath())}`);
+		log('info', `User ${chalk.green(identification!)} has been logged in with token "${name}". Credentials are stored in ${chalk.blue(getConfigFilePath())}`);
 	},
 });