+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- Respect argument `--api-url` as priority over `SOS_API_URL` environment variable and default value
+- Log info/warning output into stderr instead of stdout
+
 ## [1.1.4] - 2022-11-25
 ### Fixed
 - Removed unused `display.appcache` file from Emulator (replaced with `serviceWorker.js`)